Extract ArbitraryDataFieldView
